# docs/swagger.yml
swagger: '2.0'
info:
  version: 2.0.0
  title: XDC Network Governance DApp APIs
  description: Happy to code XDC Network Governance DApp APIs
  license:
    name: Github
    url: 'https://github.com/xinfinorg/masternodeapp'
...
schemes:
  - https
  - http
host: master.xinfin.network
consumes:
  - application/json
produces:
  - application/json
paths:
...
  '/api/transactions/candidate/{candidate}':
    parameters:
      - name: candidate
        in: path
        type: string
        description: candidate's address
        required: true
      - name: limit
        in: query
        type: number
        description: number of records
      - name: page
        in: query
        type: number
        description: page number
    get:
      tags:
        - Transaction
      summary: Get transactions of a candidate
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/txCandidate'
  '/api/transactions/voter/{voter}':
    parameters:
      - name: voter
        in: path
        type: string
        description: voter's address
        required: true
      - name: limit
        in: query
        type: number
        description: number of records
      - name: page
        in: query
        type: number
        description: page number
    get:
      tags:
        - Transaction
      summary: Get transactions of a voter
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/txVoter'
  '/api/transactions/{tx}':
    parameters:
      - name: tx
        in: path
        type: string
        description: Transaction
        required: true
    get:
      tags:
        - Transaction
      summary: Get a transaction's detail
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/transaction'
...
definitions:
...
  txCandidate:
    properties:
      items:
        type: array
        description: Candidate's transactions array
        items:
          type: object
          description: Candidate's transactions
          example:
            _id: 5c247c545e769d2f2c10ab38
            smartContractAddress: '0x0000000000000000000000000000000000000088'
            tx: '0x0f96e419c89dcf1397a6206959334a4485c5a158a0320cb8a8e98db3b7888141'
            event: Vote
            voter: '0xe91dc9746eed1b5971aabd6e1681da1a4d06be8d'
            owner: ''
            candidate: '0x11621900588eca4410c00097a9f59237f34064cd'
            capacity: '11000000000000000000'
            blockNumber: 2897395
            createdAt: '2018-12-27T07:16:36.000Z'
            __v: 0
      total:
        type: number
        description: Number of candidate
        example: 100
  txVoter:
    properties:
      items:
        type: array
        description: Candidate's transactions array
        items:
          type: object
          description: Candidate's transactions
          example:
            _id: 5c24a189ff305840a2498bf0
            smartContractAddress: '0x0000000000000000000000000000000000000088'
            tx: '0xf8f2d402c6b1cb5f891b687d69ddcd920e58d9e9fe002857cac0e7ce47998884'
            event: Unvote
            voter: '0x487d62d33467c4842c5e54eb370837e4e88bba0f'
            owner: ''
            candidate: '0xfc5571921c6d3672e13b58ea23dea534f2b35fa0'
            capacity: '10000999000000000000000'
            blockNumber: 763397
            createdAt: '2018-12-27T09:55:21.000Z'
            __v: 0
      total:
        type: number
        description: Number of candidate
        example: 100
...
  transaction:
    properties:
      _id:
        type: string
        description: id
        example: 5c52a4465f0abf3496da0e91
      tx:
        type: string
        description: transaction hash
        example: '0xdd03614f86982663b4c17eaaf9051b3df7a88fefc445df41921fa223c6178bf7'
      smartContractAddress:
        type: string
        description: smart contract's address
        example: '0x0000000000000000000000000000000000000088'
      __v:
        type: number
        description: __v
        example: 0
      capacity:
        type: string
        description: capacity in wei
        example: '1715000000000000000000'
      createdAt:
        type: string
        description: creation date of candidate
        example: '2018-10-31T03:42:39.375Z'
      owner:
        type: string
        description: owner's address
        example: '0x7e0d154cba89190ce1801031cba9d46ad21eb317'
      blockNumber:
        type: number
        description: blockNumber
        example: 1310702
      updatedAt:
        type: string
        description: update time
        example: '2019-01-11T09:17:59.535Z'
      voter:
        type: string
        description: voter address
        example: '0x7e0d154cba89190ce1801031cba9d46ad21eb317'
      candidate:
        type: string
        description: candidate address
        example: ''
      event:
        type: string
        description: event
        example: Withdraw
        enum:
          - Vote
          - Unvote
          - Withdraw
          - Propose
          - Resign
...